requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alpaca-trade-api==3.2.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
deprecation==2.1.0
exchange_calendars==4.11.1
fonttools==4.59.2
frozenlist==1.7.0
idna==3.10
kiwisolver==1.4.9
korean-lunar-calendar==0.3.1
matplotlib==3.10.6
msgpack==1.0.3
multidict==6.6.4
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
polygon-api-client==1.15.3
propcache==0.3.2
psycopg2==2.9.10
pyluach==2.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.1
requests==2.32.5
six==1.17.0
toolz==1.0.0
tqdm==4.67.1
tzdata==2025.2
urllib3==1.26.20
websocket-client==1.8.0
websockets==10.4
yarl==1.20.1
